add transaction - UAS

+++ app/Controllers/ProdukController.php
@@ -25,54 +25,80 @@ class ProdukController extends BaseController
 
     public function create()
     {
-        $dataFoto = $this->request->getFile('foto');
-
-        $dataForm = [
-            'nama' => $this->request->getPost('nama'),
-            'harga' => $this->request->getPost('harga'),
-            'jumlah' => $this->request->getPost('jumlah'),
-            'created_at' => date("Y-m-d H:i:s")
+        $rules = [
+            'nama' => 'required|min_length[6]',
+            'harga' => 'required|numeric',
+            'jumlah' => 'required|numeric'
         ];
 
-        if ($dataFoto->isValid()) {
-            $fileName = $dataFoto->getRandomName();
-            $dataForm['foto'] = $fileName;
-            $dataFoto->move('img/', $fileName);
+        if ($this->validate($rules)) {
+            $dataFoto = $this->request->getFile('foto');
+
+            $dataForm = [
+                'nama' => $this->request->getPost('nama'),
+                'harga' => $this->request->getPost('harga'),
+                'jumlah' => $this->request->getPost('jumlah'),
+                'created_at' => date("Y-m-d H:i:s")
+            ];
+
+            if ($dataFoto->isValid()) {
+                $fileName = $dataFoto->getRandomName();
+                $dataForm['foto'] = $fileName;
+                $dataFoto->move('img/', $fileName);
+            }
+
+            $this->product->insert($dataForm);
+            
+            return redirect('produk')->with('success', 'Data Berhasil Ditambah');
+
+        } else {
+            session()->setFlashdata('failed', $this->validator->listErrors());
+            return redirect()->back();
         }
 
-        $this->product->insert($dataForm);
-
-        return redirect('produk')->with('success', 'Data Berhasil Ditambah');
     } 
 
     public function edit($id)
     {
-        $dataProduk = $this->product->find($id);
-
-        $dataForm = [
-            'nama' => $this->request->getPost('nama'),
-            'harga' => $this->request->getPost('harga'),
-            'jumlah' => $this->request->getPost('jumlah'),
-            'updated_at' => date("Y-m-d H:i:s")
+        $rules = [
+            'nama' => 'required|min_length[6]',
+            'harga' => 'required|numeric',
+            'jumlah' => 'required|numeric'
         ];
 
-        if ($this->request->getPost('check') == 1) {
-            if ($dataProduk['foto'] != '' and file_exists("img/" . $dataProduk['foto'] . "")) {
-                unlink("img/" . $dataProduk['foto']);
+        if ($this->validate($rules)) {
+
+            $dataProduk = $this->product->find($id);
+
+            $dataForm = [
+                'nama' => $this->request->getPost('nama'),
+                'harga' => $this->request->getPost('harga'),
+                'jumlah' => $this->request->getPost('jumlah'),
+                'updated_at' => date("Y-m-d H:i:s")
+            ];
+
+            if ($this->request->getPost('check') == 1) {
+                if ($dataProduk['foto'] != '' and file_exists("img/" . $dataProduk['foto'] . "")) {
+                    unlink("img/" . $dataProduk['foto']);
+                }
+
+                $dataFoto = $this->request->getFile('foto');
+
+                if ($dataFoto->isValid()) {
+                    $fileName = $dataFoto->getRandomName();
+                    $dataFoto->move('img/', $fileName);
+                    $dataForm['foto'] = $fileName;
+                }
             }
 
-            $dataFoto = $this->request->getFile('foto');
+            $this->product->update($id, $dataForm);
 
-            if ($dataFoto->isValid()) {
-                $fileName = $dataFoto->getRandomName();
-                $dataFoto->move('img/', $fileName);
-                $dataForm['foto'] = $fileName;
-            }
+            return redirect('produk')->with('success', 'Data Berhasil Diubah');
+        } else {
+            session()->setFlashdata('failed', $this->validator->listErrors());
+            return redirect()->back();
         }
 
-        $this->product->update($id, $dataForm);
-
-        return redirect('produk')->with('success', 'Data Berhasil Diubah');
     }
 
     public function delete($id)
